requirements.txt
# requirements.txt
PyMuPDF
numpy
PyYAML
scipy
torch
